% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{databias}
\alias{databias}
\title{Fair Geographically Weighted Regression}
\usage{
databias(x, y)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}
}
\value{
A data frame with p or (p + 1) rows and the following columns
\itemize{
\item lower: the mean value for Y on individuals with an observed x_j value
lower than median(x_j)
\item greater: the mean value for Y on individuals with an observed x_j value
greater or equal than median(x_j)
\item u: the "u" measure of variation as returned by fair.u()
\item variation: the percentage of variation between the observations in the
lower/greater class
\item median: the median value for the variable
}
}
\description{
Show an informative table with data biases to help in the election
of a sensitive variable
}
