% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair.R
\name{fair.classes}
\alias{fair.classes}
\title{Fair Geographically Weighted Regression}
\usage{
fair.classes(x, y, d, select_col, select_val = NULL)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{select_col}{the index of the column in X containing the sensitive variable}

\item{select_val}{the value that will split observations between sensitive and
non sensitive class, usually the median is used}
}
\value{
a list with components
\itemize{
\item w.s: indices for the sensitive class observations
\item w.ns: indices for the non-sensitive class observations
\item n.s: number of observations in the sensitive class
\item n.ns: number of observations in the non-sensitive class
\item y.s: observations in the sensitive class
\item y.ns: observations in the non-sensitive class
}
}
\description{
Get which observations belong to the sensitive or non-sensitive
class in a given dataset
}
