% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair.R
\name{fair.epsilon}
\alias{fair.epsilon}
\title{Fair Geographically Weighted Regression}
\usage{
fair.epsilon(classes, fair.perc = 0)
}
\arguments{
\item{classes}{a "classes" object as returned by fair.classes()}

\item{fair.perc}{the percentage of unfairness we want to reduce}
}
\value{
the epsilon value
}
\description{
Get the epsilon value as the difference between the means of the
observed values in the sensitive and non-sensitive classes
}
