% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair.R
\name{fair.u}
\alias{fair.u}
\title{Fair Geographically Weighted Regression}
\usage{
fair.u(x, classes, betas = NULL)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{classes}{a "classes" object as returned by fair.classes()}

\item{betas}{(optional) a Beta (n x p) matrix}
}
\value{
the difference in absolute value between the
estimated means of the non-sensitive and sensitive classes
}
\description{
Get the "u" metric of variation between subjects
belonging to the sensitive or non-sensitive class
}
