% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.fair_classes}
\alias{foldcv.fair_classes}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.fair_classes(classes, y, folds)
}
\arguments{
\item{classes}{a "classes" list, as returned by fair.classes()}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{folds}{a list with the indexes of every fold, as returned by foldcv.split()}
}
\value{
a list with the sensitive/non-sensitive elements for each fold
}
\description{
Get which observations belong to the sensitive or non-sensitive
class in every fold
}
