% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.fvalue}
\alias{foldcv.fvalue}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.fvalue(x, y, d, foldgwr.result, wfun)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{foldgwr.result}{the result of k-fold CV as returned by foldcv.gwr}

\item{wfun}{weighting function}
}
\value{
an (nfolds x 2) matrix with the values for f on both test and train datasets
}
\description{
Get the objective function value when running k-fold CV
}
