% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.gwr}
\alias{foldcv.gwr}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.gwr(
  x,
  y,
  d,
  folds.h,
  wfun,
  folds,
  fair.perc = NULL,
  fold.classes = NULL,
  constrained = FALSE
)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{folds.h}{a numeric vector with the h value for every fold}

\item{wfun}{weighting function}

\item{folds}{a list with the indexes of every fold, as returned by foldcv.split()}

\item{fair.perc}{(optional) the percentage of unfairness we want to reduce}

\item{fold.classes}{(optional) a list of lists with the sensitive/non-sensitive
elements for each fold, as returned by foldcv.fair_classes()}

\item{constrained}{whether to run this as a constrained or unconstrained problem.
If FALSE, the "fold.classes" and "fair.perc" parameters will be ignored}
}
\value{
a list with a (fair) GWR solution for each fold, containing
\itemize{
\item solution: the solution Beta (n x p) matrix
\item pred: a (n x 1) matrix containing the estimated y values
\item fold: the indexes of
\item fair.perc: the requested percentage of unfairness reduction
\item bandwidth: the bandwidth used for the calculations
\item value: the value of the objective function evaluated at the solution
\item value.train: the value of the objective function for the "train" subset
\item value.test: the value of the objective function for the "test" subset
}
}
\description{
Adjust a CV (fair) GWR model
}
