% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.h_sel}
\alias{foldcv.h_sel}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.h_sel(
  x,
  y,
  d,
  wfun,
  folds,
  fair.perc = NULL,
  fold.classes = NULL,
  constrained = FALSE,
  verbose = TRUE,
  record = FALSE
)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{folds}{a list with the indexes of every fold, as returned by foldcv.split()}

\item{fair.perc}{(optional) the percentage of unfairness we want to reduce}

\item{fold.classes}{(optional) a list of lists with the sensitive/non-sensitive
elements for each fold, as returned by foldcv.fair_classes()}

\item{constrained}{whether to run this as a constrained or unconstrained problem.
If FALSE, the "fold.classes" and "fair.perc" parameters will be ignored}

\item{verbose}{show internal calculations}

\item{record}{return the points used in the internal calculations}
}
\value{
a list with the optimal h values for each fold
}
\description{
Find the optimal h value for a CV (fair) GWR model
}
