% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.mse}
\alias{foldcv.mse}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.mse(x, y, foldgwr.result, fun = NULL, type = c("test", "train"))
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{foldgwr.result}{the result of running k-fold CV fair GWR as returned by
foldcv.gwr()}

\item{fun}{after getting the MSE for every fold, which value to keep? Usually, mean
is used, but median, min or max can also be considered. Not supplying a function
only means the MSE value for every fold will be returned}

\item{type}{whether we want the MSE value for the "test" or the "train" subsets}
}
\value{
the MSE values for every fold, or a single value as calculated by "fun"
}
\description{
Get the MSE value for a CV (fair) GWR solution
}
