% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.split}
\alias{foldcv.split}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.split(x, nfolds, seed = NULL)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{nfolds}{the number of folds}

\item{seed}{(optional) set a fixed seed so results can be reproduced}
}
\value{
a list with length = folds containing the indexes pertaining to each fold
}
\description{
Get folds from a dataset
}
