% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.wmat}
\alias{foldcv.wmat}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.wmat(wmat, fold)
}
\arguments{
\item{wmat}{a weight matrix in list form, as returned by wmat.gauss() or wmat.bisquare()}

\item{fold}{a numeric vector containing the elements belonging to the "test" fold}
}
\value{
a weight matrix where the elements of the current fold have 0 weight
}
\description{
Transform a weight matrix to work with the k-fold CV problem
}
