% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldcv.R
\name{foldcv.y}
\alias{foldcv.y}
\title{Fair Geographically Weighted Regression}
\usage{
foldcv.y(y, foldgwr.result)
}
\arguments{
\item{x}{an X (n x p) design matrix}
}
\value{
TODO (4)
}
\description{
Unfinished Fold-CV documentation, subject to change
}
