% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.Amat}
\alias{gwr.Amat}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.Amat(x, classes)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{classes}{a "classes" object as returned by fair.classes()}
}
\value{
an A (n x 2) matrix
}
\description{
Generate the A matrix needed when solving
the constrained GWR problem via quadprog
}
