% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gwr.basemap}
\alias{gwr.basemap}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.basemap(geometry)
}
\arguments{
\item{geometry}{a geometry object from the original data as given by sf::st_geometry()}
}
\value{
a ggplot map
}
\description{
Generate the base map only containing the outline of the regions
}
