% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gwr.betamaps}
\alias{gwr.betamaps}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.betamaps(
  basemap,
  geometry,
  betas,
  breaks = NULL,
  titles = NULL,
  legend.titles = NULL,
  subset.i = NULL
)
}
\arguments{
\item{basemap}{the original map of the data as returned by gwr.basemap() or,
alternatively, a compatible ggplot() object}

\item{geometry}{a geometry object from the original data as given by sf::st_geometry()}

\item{betas}{a Beta (n x p) matrix}

\item{breaks}{a numeric vector containing the break points for the plotted values}

\item{titles}{a character vector of length p containing the title for the
p plots, use NULL for no title}

\item{legend.titles}{a character vector of length p containing the title for the
legend of the p plots, use NULL for no legend title}

\item{subset.i}{(optional) the subset of areas to color as an integer vector}
}
\value{
a list of length p containing a ggplot map for every beta column
}
\description{
Generate maps for a GWR Beta (n x p) matrix
}
