% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.colors}
\alias{gwr.colors}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.colors(y, breaks = NULL, highs = "Reds", lows = "Blues", reverse = FALSE)
}
\arguments{
\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{breaks}{(optional) a numeric vector containing the break points for the
plotted values}

\item{highs}{(optional) the color scale to use for values higher than 0,
"Reds" by default, or any scale available in RColorBrewer::display.brewer.all()}

\item{lows}{(optional) the color scale to use for values lower than 0,
"Blues" by default, or any scale available in RColorBrewer::display.brewer.all()}
}
\value{
a list with components
\itemize{
\item breaks_cut: a factor object containing the values to plot
\item colors: a vector containing the colors associated to
}
}
\description{
Get a color scale to be used with the plotting functions,
see 'gwr.betamaps' and 'gwr.predmap', where it's used internally
}
