% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{gwr.cvscore}
\alias{gwr.cvscore}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.cvscore(
  h,
  x,
  y,
  d,
  wfun,
  cv = TRUE,
  fair.perc = NULL,
  classes = NULL,
  fold.classes = NULL,
  constrained = FALSE
)
}
\arguments{
\item{h}{bandwidth parameter}

\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{wfun}{weighting function}

\item{cv}{use the cross-validated score function (TRUE by default)}

\item{fair.perc}{the percentage of unfairness we want to reduce}

\item{classes}{a "classes" object as returned by fair.classes()}

\item{fold.classes}{(optional) a subset of the classes parameter as returned
by fair.classes_fold(), in case we need a k-fold CV solution}

\item{constrained}{whether to run this as a constrained or unconstrained problem. If FALSE,
the "classes" and "fair.perc" parameters will be ignored}
}
\value{
the CV-score value at h
}
\description{
Return the CV-score value for the given parameters. This function
is used when optimizing the bandwidth (h) parameter, hence the need to
have it as first parameter
}
