% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.Dmat}
\alias{gwr.Dmat}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.Dmat(x, w)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{w}{a list of length n containing the weight matrices for our problem,
as returned by wmat.gauss() or wmat.bisquare()}
}
\description{
Generate the D matrix needed when solving
the GWR problem via quadprog
}
