% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.dvec}
\alias{gwr.dvec}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.dvec(x, y, w)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{w}{a list of length n containing the weight matrices for our problem,
as returned by wmat.gauss() or wmat.bisquare()}
}
\description{
Generate the d vector needed when solving
the GWR problem via quadprog
}
