% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.fvalue2}
\alias{gwr.fvalue2}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.fvalue2(x, y, w, betas)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{w}{a list of length n containing the weight matrices for our problem,
as returned by wmat.gauss() or wmat.bisquare()}
}
\value{
the value for f
}
\description{
Get the objective function value for a GWR problem
}
