% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{gwr.h_sel}
\alias{gwr.h_sel}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.h_sel(
  x,
  y,
  d,
  wfun,
  fair.perc = NULL,
  classes = NULL,
  fold.classes = NULL,
  constrained = FALSE,
  verbose = TRUE,
  progress = FALSE,
  shiny.progress = NULL,
  record = FALSE
)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{wfun}{weighting function}

\item{verbose}{show internal calculations}

\item{progress}{like "verbose", but prettier, can't use when "verbose" = TRUE}

\item{shiny.progress}{a "progress" object as given by shiny::Progress$new(), to use
with a shiny webapp}

\item{record}{return the points used in the internal calculations}
}
\value{
a list with components
\itemize{
\item h: optimal bandwidth
\item value: CV-score value at h
\item record: (optional) matrix with h and CV score values used
}
}
\description{
Select the optimal bandwidth on a given dataset
}
