% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gwr.highlightmap}
\alias{gwr.highlightmap}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.highlightmap(basemap, geometry, subsets, fill.colors)
}
\arguments{
\item{basemap}{the original map of the data as returned by gwr.basemap() or,
alternatively, a compatible ggplot() object}

\item{geometry}{a geometry object as given by sf::st_geometry()}

\item{subsets}{the indexes of the subsets to be plotted, as a list
of integer vectors}

\item{colors}{the color for every subset, as a character vector}
}
\value{
the plotted map with colored predictions per tract
}
\description{
Generate a map highlighting the specified areas, useful e.g.
when illustrating methods such as k-fold CV
}
\examples{
# Example for two colors on some data
subsets <- list(c(1,2,3), c(4,5,6))
fill.colors <- c("lightgrey", "lightyellow")
gwr.highlightmap(mydata.basemap,
                 mydata$geometry,
                 subsets,
                 fill.colors)
}
