% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gwr.names}
\alias{gwr.names}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.names(p, intercept = TRUE)
}
\arguments{
\item{p}{the number of columns of our Beta (n x p) matrix (usually returned
by gwr())}

\item{intercept}{whether the first column is the intercept (beta_0) term}
}
\value{
a character vector containing the new column names
}
\description{
Simple helper function returning the names for the Beta matrix
}
