% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gwr.predmap}
\alias{gwr.predmap}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.predmap(
  basemap,
  geometry,
  y,
  breaks = NULL,
  title = NULL,
  legend.title = NULL,
  color.scales = NULL,
  color.rev = FALSE,
  subset.i = NULL
)
}
\arguments{
\item{basemap}{the original map of the data as returned by gwr.basemap() or,
alternatively, a compatible ggplot() object}

\item{geometry}{a geometry object as given by sf::st_geometry()}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{breaks}{a numeric vector containing the break points for the plotted values}

\item{title}{(optional) the plot title}

\item{legend.title}{(optional) the legend title}

\item{color.scales}{(optional) a list containing a "highs" and "lows" value for the
colors to use in the plot, as needed by gwr.colors()}

\item{color.rev}{whether to use inversed colors (FALSE by default)}

\item{subset.i}{(optional) the subset of areas to color as an integer vector}
}
\value{
the plotted map with colored predictions per tract
}
\description{
Generate a colored prediction map
}
