% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.R
\name{gwr}
\alias{gwr}
\title{Fair Geographically Weighted Regression}
\usage{
gwr(
  x,
  y,
  d,
  h,
  wfun = NULL,
  wmat = NULL,
  fair.perc = NULL,
  classes = NULL,
  fold.classes = NULL,
  constrained = FALSE
)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{h}{bandwidth parameter}

\item{wfun}{weighting function}

\item{wmat}{(optional) use an alternate extended weight matrix, such as when
doing cross-validation or k-fold cross-validation}

\item{fair.perc}{(optional) the percentage of unfairness we want to reduce}

\item{classes}{(optional) a "classes" object as returned by fair.classes()}

\item{fold.classes}{(optional) a subset of the classes parameter as returned
by fair.classes_fold(), in case we need a k-fold CV solution}

\item{constrained}{whether to run this as a constrained or unconstrained problem. If FALSE,
the "classes" and "fair.perc" parameters will be ignored}

\item{centroids}{(optional) ...}
}
\value{
a list with components
\itemize{
\item solution: the solution Beta (n x p) matrix
\item value: the value of the objective function evaluated at the solution
\item pred: a (n x 1) matrix containing the estimated y values
\item unconstrained.solution: the unconstrained solution Beta (n x p) matrix
\item unconstrained.pred: a (n x 1) matrix containing the estimated y values for
with the unconstrained solution
\item bandwidth: the bandwidth used for the calculations
}
}
\description{
Solve the (constrained) GWR problem using a global approach
}
