% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.R
\name{gwr.simple}
\alias{gwr.simple}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.simple(x, y, d, h, wfun, verbose = FALSE)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{h}{bandwidth parameter}

\item{wfun}{weighting function}

\item{verbose}{show progress while running}
}
\value{
a list with components
\itemize{
\item solution: the solution Beta (n x p) matrix
\item value: the value of the objective function evaluated at the solution
\item pred: a (n x 1) matrix containing the estimated y values
\item bandwidth: the bandwidth used for the calculations
}
}
\description{
Solve the GWR problem using an iterative approach
}
