% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.R
\name{gwr.simple_i}
\alias{gwr.simple_i}
\title{Fair Geographically Weighted Regression}
\usage{
gwr.simple_i(
  x,
  y,
  d,
  h,
  i,
  wfun,
  cv = FALSE,
  Amat = NULL,
  bvec = NULL,
  constrained = FALSE
)
}
\arguments{
\item{x}{an X (n x p) design matrix}

\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{d}{a distance matrix (n x n)}

\item{h}{bandwidth parameter}

\item{i}{the i-th location in which we want to run this function}

\item{wfun}{weighting function}

\item{cv}{use cross-validation (weight of i is 0)}

\item{Amat}{the A matrix to use in the constrained problem}

\item{bvec}{the b vector to use in the constrained problem}

\item{constrained}{whether this is a constrained or unconstrained problem}
}
\value{
a list with components
\itemize{
\item solution: the solution beta vector
\item value: the value of the objective function evaluated at the solution
\item pred: a (n x 1) matrix containing the estimated y values
\item bandwidth: the bandwidth used for the calculations
\item unconstrained.solution: the solution beta vector for the
unconstrained problem
\item unconstrained.pred: a (n x 1) matrix containing the estimated y values
for the unconstrained problem
}
}
\description{
Solve the GWR problem for a unique location
}
