% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{medse}
\alias{medse}
\title{Fair Geographically Weighted Regression}
\usage{
medse(y, yhat)
}
\arguments{
\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{yhat}{a vector of length n or a (n x 1) matrix containing the estimated values}
}
\value{
the median SE value
}
\description{
Get the median squared error (MSE) for a GWR prediction compared to
the original y values
}
