% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mse}
\alias{mse}
\title{Fair Geographically Weighted Regression}
\usage{
mse(y, yhat)
}
\arguments{
\item{y}{a vector of length n or a (n x 1) matrix containing the observed values}

\item{yhat}{a vector of length n or a (n x 1) matrix containing the estimated values}
}
\value{
the MSE value
}
\description{
Get the mean squared error (MSE) for a GWR prediction compared to
the original y values
}
