% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim.getall}
\alias{sim.getall}
\title{Fair Geographically Weighted Regression}
\usage{
sim.getall(n, p)
}
\arguments{
\item{n}{the number of elements}

\item{p}{the number of variables (first one being the Intercept variable)}
}
\value{
a list with components
\itemize{
\item data: an st_sfc object containing the simulated data
\item dist: the distance matrix for the simulated data
\item h: the used bandwidth
\item w: the used weight matrices
\item betas: the Beta (n x p) coefficients matrix from which y is generated
\item residuals: residuals used to estimate y, generated from a MVN(0, sigma)
\item sigma: the sigma matrices used to generate the residuals, in vector form
}
}
\description{
Meta-function returning a simulated dataset of n points
}
\references{
Harris, P., Fotheringham, A. S., Crespo, R., & Charlton, M. (2010). The use of geographically weighted regression for spatial prediction: an evaluation of models using simulated data sets. Mathematical Geosciences, 42, 657-680.
}
