% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim.getbetas}
\alias{sim.getbetas}
\title{Fair Geographically Weighted Regression}
\usage{
sim.getbetas(x, coords)
}
\arguments{
\item{x}{an X (n x p) design matrix}
}
\value{
a Beta (n x p) matrix
}
\description{
Return the Beta matrix for a simulated dataset
}
