% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim.getcoords}
\alias{sim.getcoords}
\title{Fair Geographically Weighted Regression}
\usage{
sim.getcoords(n)
}
\arguments{
\item{n}{the number of elements}
}
\value{
an X (n x 2) matrix
}
\description{
Simulate the coordinates for n points
}
