% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim.getx}
\alias{sim.getx}
\title{Fair Geographically Weighted Regression}
\usage{
sim.getx(n, p)
}
\arguments{
\item{n}{the number of elements}
}
\value{
an X (n x 2) matrix
}
\description{
Simulate an X (n x 2) matrix
}
