% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{wmat.bisquare}
\alias{wmat.bisquare}
\title{Fair Geographically Weighted Regression}
\usage{
wmat.bisquare(d, h, i = NULL)
}
\arguments{
\item{d}{a distance matrix (n x n)}

\item{h}{bandwidth parameter}

\item{i}{the i-th observation for which we need a weight matrix. If NULL (default),
all the observations are considered instead}
}
\value{
a list of n weight matrices, or a single W_i matrix, in vector (non-diagonal) form
}
\description{
Generate a matrix or list of matrices containing bi-squared weights
of the form {1 - (d/h)^2}^2. When the distance between two points is smaller
than the bandwidth, 0 is returned
}
