% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{wmat.gauss}
\alias{wmat.gauss}
\title{Fair Geographically Weighted Regression}
\usage{
wmat.gauss(d, h, i = NULL)
}
\arguments{
\item{d}{a distance matrix (n x n)}

\item{h}{bandwidth parameter}

\item{i}{the i-th observation for which we need a weight matrix. If NULL (default),
all the observations are considered instead}
}
\value{
a list of n weight matrices, or a single W_i matrix, in vector (non-diagonal) form
}
\description{
Generate a matrix or list of matrices containing gaussian weights
of the form exp{(-1/2) * (d / h) ^ 2}. When the distance between two
points is smaller than the bandwidth, 0 is returned
}
